+++ lib/src/corner_banner.dart
