+++ lib/super_banners.dart
@@ -1,5 +1,5 @@
 library super_banners;
 
-export 'src/banner.dart';
-export 'src/built_with_flutter_banner.dart';
-export 'src/positioned_banner.dart';
+export 'src/corner_banner.dart';
+export 'src/built_with_flutter_corner_banner.dart';
+export 'src/positioned_corner_banner.dart';
